make path function moved into path_utils
